Implement Module to control the LEDPwm Peripheral
This commit includes the base device tree for the DE1-SoC Board
(socfpga_cyclone5_de1soc.dts) as well as an addition, that extends that
device tree by the PWM Controlled LEDs (socfpga_cyclone5_de1soc_ldd.dts)

Each LED will be controlled by a character device, /dev/ledN where N is
the number of the LED.
Note: this driver enumerates the LEDs based on the order they are
registered, which may not correspond to the layout on the Hardware.

A LED can be controlled by writing a 0-100 Byte to the corresponding
character device, this controlls the brightness in Percent.
The device can also be read, which provides the current brightness.

Signed-off-by: Alexander Daum <alexander.daum@mailbox.org>