Add a pushbutton driver for the DE1_SOC keys
When a button is released, this driver stores the button state before
release ('1' means a button was pressed, '0' means it was not pressed).
The events are stored in a 8 Long FIFO.

The driver provides one read-only character device pushbutton_event to
read the events. Reading from the device will return one byte for each
event in the fifo, where one bit represents the state of one button
(bit0 = Key0).

If no events are currently in the FIFO, the read operation will block
until a Button is released.

Signed-off-by: Alexander Daum <alexander.daum@mailbox.org>